earth orbiting, but going further than 1 UA

+++ src/utils/Constants.java
@@ -6,9 +6,9 @@ public class Constants {
     public static final int WIDTH = 1280;
     public static final int HEIGHT = 720;
     
-    public static final float UA = 1.496e11f;
-    public static final float SCALE = 250.0f / UA; 
+    public static final double UA = 1.496e11;
+    public static final double SCALE = 250.0f / UA; 
     
     public static final float DELTA = 3600*24;
-    public static final float G = 6.67430e-11f;
+    public static final double G = 6.67430e-11f;
 }
